requirements.txt
fastapi
uvicorn
python-dotenv
pandas
pydantic